Ai_model/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.7
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0.1
tqdm>=4.66.0
opencv-python>=4.9.0.80
Pillow>=10.3.0
matplotlib>=3.8.0

